requirements.txt
aioredis==1.3.1
amqp==2.5.2
appdirs==1.4.3
appnope==0.1.0
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
backcall==0.1.0
billiard==3.6.3.0
CacheControl==0.12.6
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
daphne==2.5.0
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.0.6
django-celery-results==1.2.1
django-environ==0.4.5
django-filter==2.2.0
django-filters==0.2.1
django-redis==4.11.0
djangorestframework==3.11.0
greenlet==0.4.15
gunicorn==20.0.4
hiredis==1.0.1
html5lib==1.0.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
inflection==0.4.0
ipaddr==2.2.0
ipdb==0.11
ipython==7.3.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
kombu==4.6.8
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
nltk==3.5
numpy==1.18.4
packaging==20.3
parso==0.7.0
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
progress==1.5
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
pynvim==0.4.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
PyYAML==5.3.1
redis==3.5.2
redisai==1.0.1
regex==2020.5.14
requests==2.23.0
retrying==1.3.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
tqdm==4.46.0
traitlets==4.3.3
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
wcwidth==0.1.9
webencodings==0.5.1
whitenoise==5.1.0
zipp==3.1.0
zope.interface==5.1.0
